+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to **Pipecat Flows** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [TBD]
+
+### Deprecated
+
+- Static Flows are now deprecated and will be removed in v1.0.0. Use Dynamic
+  Flows in their place. The deprecation includes the `flow_config` arg of the
+  `FlowManager` and the `FlowConfig` type.
+
 ## [0.0.18] - 2025-06-27
 
 ### Added